% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_info}
\alias{cell_info}
\title{Cell line information}
\format{
## `cell_info`
A data frame with 100 rows and 5 columns:
\describe{
  \item{ModelID}{Cell line DepMap ID}
  \item{cell}{Tissue of the cell line, original name in DepMap is OncotreeLineage}
  \item{origin_net}{Name of original metabolic network downloaded from Metabolic Atlas}
  \item{net}{Name of transformed enzyme network by Met2Graph package}
}
}
\source{
DepMap
}
\usage{
cell_info
}
\description{
A subset data, which contains 100 cell lines of 5 tissues, of DepMap data.
}
\keyword{datasets}
