% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpg_gene}
\alias{cpg_gene}
\title{Gene CPG features}
\format{
## `cpg_gene`
A data frame with 3247 rows and 3075 columns:
\describe{
  \item{A4GALT}{The CPG feature of gene A4GALT}
  ...
}
}
\source{
MSigDB database
}
\usage{
cpg_gene
}
\description{
To obtain the features of genes, we collected the CGP (chemical and genetic perturbations) gene sets from MSigDB database.
After filter gene sets without any gene existed in GSMs, we got 3247 gene sets.
For each gene to be predicted, its feature is a 3247-dimensional 0/1 vector
}
\keyword{datasets}
