% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dep_data}
\alias{dep_data}
\title{Gene dependency data}
\format{
## `dep_data`
A data frame with 436325 rows and 4 columns:
\describe{
  \item{ModelID}{Cell line DepMap ID}
  \item{gene}{Gene symbol}
  \item{score}{post-Chronos gene dependency probability score}
  \item{type}{Dependency label, when score > 0.5 the label is 1, otherwise 0}
}
}
\source{
DepMap database
}
\usage{
dep_data
}
\description{
Subset of post-Chronos gene dependency probability score data obtained from Depmap.
This subset data only contains 25 skin cancer cell lines.
}
\keyword{datasets}
