% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensg2name}
\alias{ensg2name}
\title{Convert ENSEMBL ID to Gene Symbol and get corresponding features}
\usage{
ensg2name(ensg, mapping, exp_list = NULL, pathway)
}
\arguments{
\item{ensg}{A character String, indicting ENSEMBL gene ID; gene complex with multiple gene ID separated by commas}

\item{mapping}{A dataframe, gene ID mapping file, containing at least two columns of `ensembl_id` and `symbol`}

\item{exp_list}{A character vector, containing expressed gene symbols.}

\item{pathway}{A dataframe, gene CPG features, which rowname is name of pathway and colname is gene symbol}
}
\value{
A list with three elements, gene symbol, gene CPG feature, and whether the gene is expressed
}
\description{
Convert ENSEMBL gene ID to Gene symbol, extract CPG gene features and gene expression information.
}
\examples{
library(dplyr)
data("enz_gene_mapping")
data("cpg_gene")
data("SkinCancerNet")
data("cell_info")
data("gene_exp")
cell_info <- cell_info \%>\% filter(grepl("SkinCancer",origin_net))
gene_exp <- gene_exp \%>\% select(any_of(cell_info$ModelID))

##Net with gene expressed in cell ACH-001521
cell_exp <- gene_exp["ACH-001521"]
cell_exp$gene <- rownames(cell_exp)
colnames(cell_exp)[1] <- "exp"
cell_exp <- cell_exp$gene[which(cell_exp$exp > 1)]

SkinCancerNet <- SkinCancerNet[1:10,]
cell_net <- data.frame(id = unique(c(SkinCancerNet$from,SkinCancerNet$to))) \%>\%
   rowwise() \%>\%
   mutate(gene = ensg2name(id,enz_gene_mapping,cell_exp,cpg_gene)[[1]],
          fea = ensg2name(id,enz_gene_mapping,cell_exp,cpg_gene)[[2]],
          is_exp = ensg2name(id,enz_gene_mapping,cell_exp,cpg_gene)[[3]]) \%>\%
   ungroup()

}
