% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enz_gene_mapping}
\alias{enz_gene_mapping}
\title{Gene ID mapping file}
\format{
## `enz_gene_mapping`
A data frame with 40102 rows and 3 columns:
\describe{
  \item{gene_id}{Entrez gene ID}
  \item{ensembl_id}{ENSEMBL gene ID}
  \item{symbol}{Gene Symbol}
}
}
\usage{
enz_gene_mapping
}
\description{
Gene ID mapping between Entrez gene ID and ENSEMBL ID and Gene Symbol
}
\keyword{datasets}
