% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_exp}
\alias{gene_exp}
\title{Gene expression of cell lines}
\format{
## `gene_exp`
A data frame with 19193 rows and 81 columns:
\describe{
  \item{ACH-001521}{Gene mRNA expression of cell line ACH-001521, unit is log2(TPM+1)}
  ...
}
}
\source{
DepMap
}
\usage{
gene_exp
}
\description{
A subset data, which contains gene expression of 81 cell lines , of DepMap data.
}
\keyword{datasets}
