% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dep}
\alias{get_dep}
\title{Get dependency probability label for gene}
\usage{
get_dep(gene, dep_data, per_cell_mode = FALSE, cell_name = NULL)
}
\arguments{
\item{gene}{Character string, gene name or gene complex with multiple gene names separated by commas.}

\item{dep_data}{Dataframe, dependency probability score data with at least three columns, ModelID, gene and type.}

\item{per_cell_mode}{Boolean value, whether in single cell mode}

\item{cell_name}{Character string, only used when `per_cell_mode` is TRUE, cell line ID, match with ModelID in `dep_data`}
}
\value{
Dependency probability lable for specific gene
}
\description{
For one specific gene or gene complex, extract corresponding dependency probability score.
The dependency probability score of the gene complex is the max score of genes in the gene complex.
}
\examples{
data("dep_data")
get_dep(gene = "ACAA1,EHHADH,HSD17B4", dep_data = dep_data, per_cell_mode = TRUE, cell_name = "ACH-000014")
}
