% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gtex_exp5}
\alias{Gtex_exp5}
\title{Gene expression data of normal tissues from GTEx database}
\format{
## `Gtex_exp5`
A data frame with 56156 rows and 6 columns:
\describe{
  \item{Description}{Gene symbol}
  \item{Bladder}{Gene expression of bladder tissue, the expression is median expression of bladder samples, in TPM units}
  ...
}
}
\source{
GTEx database
}
\usage{
Gtex_exp5
}
\description{
Subset of gene expression data of GTEx database normal tissue. Only select five tissues.
}
\keyword{datasets}
