% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_gene_order}
\alias{model_gene_order}
\title{Genes used in input of DeepMeta}
\format{
## `model_gene_order`
A character vector of length 7993
}
\source{
GTEx database
}
\usage{
model_gene_order
}
\description{
The order of genes in the differential expression input of DeepMeta
}
\keyword{datasets}
