% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PancreaticCancerNet}
\alias{PancreaticCancerNet}
\title{Edge list of enzyme network of Pancreatic Cancer}
\format{
## `PancreaticCancerNet`
A data frame with 21500 rows and 3 columns:
\describe{
  \item{from}{One node}
  \item{to}{Another node}
  \item{Metabolite}{Metabolite ID which connect from node and to node}
}
}
\usage{
PancreaticCancerNet
}
\description{
Cancer-specific enzyme network represented by edge list, converted from XML file using Met2Graph
}
\keyword{datasets}
