% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreDeepMeta-package.R
\docType{package}
\name{PreDeepMeta-package}
\alias{PreDeepMeta}
\alias{PreDeepMeta-package}
\title{PreDeepMeta: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
