% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreDiffExp.R
\name{PreDiffExp}
\alias{PreDiffExp}
\title{Prepare DeepMeta model differential expression profile input}
\usage{
PreDiffExp(
  tumor_exp,
  normal_exp,
  tumor_normal_mapping,
  gene_order,
  save_file = FALSE,
  save_path = NULL
)
}
\arguments{
\item{tumor_exp}{A dataframe, gene expression of tumor samples, rows are gene names and cols are sample names, unit is log2(TPM+1)}

\item{normal_exp}{A dataframe, normal tissue median expression, the first column with name `Description` is gene symbol and following columns are different tissues, unit is TPM}

\item{tumor_normal_mapping}{A dataframe, mapping between sample name (in `tumor_exp`) and tissue name (in `normal_exp`)}

\item{gene_order}{A character vector, containing the genes of DeepMeta input.}

\item{save_file}{A boolean value, whether save file}

\item{save_path}{A character string, file save path, needed when `save_file` is TRUE}
}
\value{
A dataframe of differential expression with first column is sample name and following columns are gene names
}
\description{
DeepMeta framework has two inputs, the sample-specific enzyme network and the differential expression data.
The sample-specific enzyme network (SEN) is constructed from cancer-specific or tissue-specific GSMs (Genome-Scale Metabolic Models),
and in this network, enzymes are the nodes connected by edges represented by metabolites.
SEN are constructed by removing genes in the cancer-specific enzyme network that are not expressed in the sample,
defined as expression levels below 1 TPM. The features of gene nodes in the enzyme network are composed of 3247-dimensional binary (0/1) vector,
indicating whether the gene is exist in the collected 3247 CPG (chemical and genetic perturbation) gene set. This function
prepares gene differential expression profile input, consisting 7993 gene differential expression compared to GTEx normal tissue.
}
\examples{
library(dplyr)
data("gene_exp")
data("Gtex_exp5")
data("cell_info")
data("model_gene_order")
res <- PreDiffExp(tumor_exp = gene_exp, normal_exp = Gtex_exp5,
tumor_normal_mapping = cell_info, gene_order = model_gene_order,
save_file = FALSE)
}
