% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreEnzymeNet.R
\name{PreEnzymeNet}
\alias{PreEnzymeNet}
\title{Prepare DeepMeta model sample-specific enzyme network input}
\usage{
PreEnzymeNet(
  gene_exp,
  network,
  gene_mapping,
  gene_feature,
  cell_name,
  dep_data = NULL,
  save_path
)
}
\arguments{
\item{gene_exp}{A dataframe, rows are gene and cols are cell line or sample name, unit is log2(TPM + 1)}

\item{network}{A dataframe, network in edgelist format, at least three columns with name of `from` and `to`, indicating nodes}

\item{gene_mapping}{A dataframe, mapping of different gene IDs, at least contains two columns with name of `ensembl_id` and `symbol`}

\item{gene_feature}{A dataframe, gene CPG features, rows are CPG pathways, cols are gene symbol}

\item{cell_name}{A character string, cell line or sample name, must match name in `gene_exp` columns and `ModelID` column in `dep_data`}

\item{dep_data}{A dataframe, gene dependency data, at least contains three columns with name of `ModelID`, `gene` and `type`}

\item{save_path}{A character string, path name of output file saved}
}
\value{
Two txt files, `.txt` file is the edge list format of SEN and `_feat.txt` file contains gene CPG features
}
\description{
DeepMeta framework has two inputs, the sample-specific enzyme network and the differential expression data.
The sample-specific enzyme network (SEN) is constructed from cancer-specific or tissue-specific GSMs (Genome-Scale Metabolic Models),
and in this network, enzymes are the nodes connected by edges represented by metabolites.
SEN are constructed by removing genes in the cancer-specific enzyme network that are not expressed in the sample,
defined as expression levels below 1 TPM. The features of gene nodes in the enzyme network are composed of 3247-dimensional binary (0/1) vector,
indicating whether the gene is exist in the collected 3247 CPG (chemical and genetic perturbation) gene set. This function
prepare the network input consisting of SEN (.txt file) and gene features (_feat.txt file).
}
\examples{
library(dplyr)
data("gene_exp")
data("SkinCancerNet")
data("enz_gene_mapping")
data("cpg_gene")
data("dep_data")
PreEnzymeNet(gene_exp, network = SkinCancerNet,
gene_mapping = enz_gene_mapping, gene_feature = cpg_gene,
cell_name = "ACH-001521", dep_data = dep_data, save_path = tempdir())
##or without dependency data
PreEnzymeNet(gene_exp, network = SkinCancerNet,
gene_mapping = enz_gene_mapping, gene_feature = cpg_gene,
cell_name = "ACH-001521", save_path = tempdir())
}
