% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SkinCancerNet}
\alias{SkinCancerNet}
\title{Edge list of enzyme network of Skin Cancer}
\format{
## `SkinCancerNet`
A data frame with 18441 rows and 3 columns:
\describe{
  \item{from}{One node}
  \item{to}{Another node}
  \item{Metabolite}{Metabolite ID which connect from node and to node}
}
}
\usage{
SkinCancerNet
}
\description{
Cancer-specific enzyme network represented by edge list, converted from XML file using Met2Graph
}
\keyword{datasets}
