% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UrothelialCancerNet}
\alias{UrothelialCancerNet}
\title{Edge list of enzyme network of Urothelial Cancer}
\format{
## `UrothelialCancerNet`
A data frame with 18553 rows and 3 columns:
\describe{
  \item{from}{One node}
  \item{to}{Another node}
  \item{Metabolite}{Metabolite ID which connect from node and to node}
}
}
\usage{
UrothelialCancerNet
}
\description{
Cancer-specific enzyme network represented by edge list, converted from XML file using Met2Graph
}
\keyword{datasets}
